+++ files/most_current.py
@@ -1,38 +1,26 @@
-#updated section 9/24/19 6:00PM
+#put save states here: 
+#while making more edits on most current
+#here will be the last fully working most_current.py file
+
+#updated section 3/20/20 8:10PM
 #anichau and jeffchen
 #signed jeffrey chen
-
+#make the cars spawn closer together with less in
 import pygame
 import random
-import math
-import sys
-import tkinter
+import math  
+import os.path
+import time
+from time import gmtime, strftime
 
 '''
 notes:
 YELLOW CARS = HAVE DROPPED OFF ALREADY
-deceleration function?
 
 next step: 
-exiting cars range from 1 to 3 passengers
-count number of student flow?
-can definitely start doing actual statistics
-count number of cars that got deleted in x amount of time = flow
 do average speed, etc. in mph
 
-Exiting students should be blue, roughly same circle size
-travel about 50 pix
-work on passenger
-
-drop off list of times to wait - access that through the rand int
-
-PASSENGERS ARE STILL NOT WORKING - CAR NEEDS TO BE ABLE TO WAIT
-2 options - have make_passenger inside the main loop make it however many times OR
-have make_passenger in def function have the number of passengers to make
-
-output things to a file to make more realisitc - at this point in time, make timestamp and write a line about car
-
-future plans: film directly translate to simulation - read in and reflect (through drone)
+COMMON LOCATIONS
 '''
 
 # Define colors
@@ -44,61 +32,47 @@ BLUE = (0,0,255)
 YELLOW = (255, 255, 0)
 
 #define other variables 
+SCREEN_WIDTH = 1430
+SCREEN_HEIGHT = 850
 CAR_SIZE = 15
-SPEED = 5
+SPEED = 0.1
 PERSON_SPEED = 1.5
 PERSON_WALK = 200
-MAXSPEED = 2
-#ACCEL = 0.03
-ACCEL = 1
-RANDOMPARAM = 10
+MAXSPEED = 1.5
+ACCEL = 0.5
+RANDOMPARAM = 8
 INFRONT = 35
 SWEEP = 20
 
-# wait time is currently unncessary
-#WAITTIME = 10000
+WAITTIME = 100000
 #change dt for acceleration change, bigger = faster accel
-#dt = 0.004
-dt = 1
-car_list = []
-person_list = []
-passenger_list = []
+dt = 0.0006
 
-#get screen size with tkinter
-root = tkinter.Tk()
-screen_width = root.winfo_screenwidth()
-screen_height = root.winfo_screenheight()
-print(screen_height)
-#print(screen_width)
-
-#subtract on the height because there are borders and everything
-size = [screen_width, int(screen_height*0.95)]
+#set height and width of screen
+size = [SCREEN_WIDTH, SCREEN_HEIGHT]
 screen = pygame.display.set_mode(size)
-
 
 class Car:
     """
     Class to keep track of a car's location and speed
     """
     def __init__(self):
-        self.x = screen_width *0.865
-        self.y = screen_height * 0.75
-        #testing for self.y = 300
+        self.x = 1250
+        self.y = 700
+        #below is spawn cars close to bottom
+        #self.x = 150
+        #self.y = 600
         self.speed = SPEED/2
         self.accel = 0
         self.color = RED
-        self.passenger_num = random.randint(2,4)
-        self.passenger_check = 0
-
-        '''
-        self.waiting_time = 0
-        self.now_time = 0
-        '''
 
         self.drop_val = -1
         self.drop_x = 0
         self.drop_y = 0
+
+        self.drop_bool = True
         self.accel_bool = True
+        self.is_dropped = False
         self.pass_bool = True
 
 
@@ -160,16 +134,6 @@ class Car:
 
         self.accel+= ACCEL
 
-class Passenger:
-    def __init__(self):
-        self.x = 0
-        self.y = 0
-        self.initial = -1
-        self.passenger_value = -1
-        self.speed = PERSON_SPEED
-        self.color = BLUE
-        self.top = True
-
 class Person:
     """
     Class to make person randomly spawn in 3+ randomly
@@ -181,6 +145,7 @@ class Person:
         self.speed = PERSON_SPEED
         self.initial = 0
         self.personal_value = random.randint(0,2)
+        #self.personal_value = 1
    
     def stop(self):
         self.speed = 0
@@ -190,14 +155,7 @@ def make_car():
     Function to make a new, random car.
     """
     car = Car()                                                                      
-    car.drop_val = random.randint(0,1)
-    #car.drop_val = 0
-
-    if car.drop_val == 0:
-        car.drop_x = random.randint(int(screen_width*.14), int(screen_width*.5))
-
-    elif car.drop_val == 1:
-        car.drop_y = random.randint(int(screen_height*0.17), int(screen_height)*0.6)
+    car.drop = random.randint(77, 250) * 2
 
     return car
 
@@ -208,64 +166,53 @@ def make_person():
 
     #right side
     if person.personal_value == 0:
-        person.x = screen_width*0.79
+        #pass
+        person.x = 1130
         person.initial = person.x
-        person.y = screen_height*0.4
+        person.y = 400
 
     #top level
     elif person.personal_value == 1:
-        #person.x = 800
-        person.x = screen_width * 0.4
-        person.y = screen_height*0.3
+        person.x = 800
+        person.y = 230
         person.initial = person.y
 
     #left side
     elif person.personal_value == 2:
-        person.x = screen_width*0.2
+        person.x = 260
         person.initial = person.x
-        person.y = screen_height*0.5
+        person.y = 450
 
     return person
- 
-
-#make_passenger(car.drop_x, car.y -15)
-def make_passenger(x, y):
-    passenger = Passenger()
-    passenger.x = x
-    passenger.y = y
-    passenger.initial = y
-    passenger.top = True
-    passenger_list.append(passenger)
-
-    return passenger
-
-    #make passenger students that exit from the dropped off cars
  
 def main():
     '''
     This is our main program.
     '''
-    #start program and give name
     pygame.init()
     pygame.display.set_caption("PHS Car Traffic")
  
     # Loop until the user clicks the close button.
     done = False
+    hundred_cars_bool = False
  
     # Used to manage how fast the screen updates
     clock = pygame.time.Clock()
  
-    wait_list = [10,200000,30000000]
+    car_list = []
+    person_list = []
+    car_count = 0
 
     car = make_car()
     car_list.append(car)
-    first_cross = make_person()
-    person_list.append(first_cross)
+    #first_cross = make_person()
+    #person_list.append(first_cross)
+    my_seconds = 0
  
     # -------- Main Program Loop -----------
     while not done:
         # --- Event Processing
-
+    
         for event in pygame.event.get():
 
             #on sublime, use below    
@@ -277,7 +224,8 @@ def main():
 
                 for width in range(0, 100):
                     for height in range(0, 100):
-                        spawn_value = screen.get_at(( int(screen_width*0.835) , int(screen_height*0.75) ))
+                        spawn_value = screen.get_at(( (1200 + width) , (650 + height) ))
+                        
                         if spawn_value == RED:
                             spawn_bool = False
                             break
@@ -308,114 +256,99 @@ def main():
                     person_list.append(person)
             '''
         
-        
         #random car spawn
         random_car = random.randint(1,RANDOMPARAM)
 
         #view finding of rectangle to make sure nearest car is out of area
         spawn_bool = True
         for width in range(0, 100):
-            for height in range(0, 100):
+            for height in range(0, 50):
                 spawn_value = screen.get_at(( (1200 + width) , (650 + height) ))
                 if spawn_value == RED:
                     spawn_bool = False
                     break
-
 
         #check if random match and there is a car within rectangular area of start
         if (random_car==3) and spawn_bool:
         #and len(car_list)<4: - len car list to restrict car number at once on screen
             car = make_car()
             car_list.append(car)
-        
 
         
         #random pedestrian spawn
-        random_ped = random.randint(1, RANDOMPARAM+30)
+        random_ped = random.randint(1, RANDOMPARAM+100)
         #random_ped = 3
         if (random_ped == 3):
-            person = make_person()
-            person_list.append(person)
+            pass
+            #see how this goes 
+            #person = make_person()
+            #person_list.append(person)
         
  
         # --- car logic
         for car in car_list:
             #Move the car's center, check for position to move car where
             car.accel_bool = True
-            car.pass_bool = True
+            
 
-            car.waiting_time = pygame.time.get_ticks()
-
-            #print(passenger_list)
-            #cooldown?
-
-            #print(car.waiting_time)
             #see if car is on right side, will go vertical up
-
-            turn_left_val = 0.19
-            if car.x > screen_width/2 and car.y > screen_height*turn_left_val:
+            if car.x > 155 and car.y > 100:
                 car.detect(0)
                 if car.accel_bool:
                     car.accelerate(0)
-               
 
 
             #see if car is on top side, will go horizontal left 
-            elif car.x > screen_width*0.13 and car.y < screen_height*turn_left_val:
-
-                if car.drop_val ==0:
-                    #while(int(car.x)==car.drop_x) or (int(car.x-1)==car.drop_x) or (int(car.x+1)==car.drop_x):  
-                    if(int(car.x)==car.drop_x) or (int(car.x-1)==car.drop_x) or (int(car.x+1)==car.drop_x):  
+            
+            elif car.x > 155 and car.y < 100:
+                the_wait_list = [0.005, 0.075, 0.01]
+                if(int(car.x)==car.drop_x) or (int(car.x-1)==car.drop_x): 
+                    car.stop(True)
+                    car.x -= the_wait_list[random.randint(0,2)]
+                    '''
+                    #while small_timer < 5:         
+                    for x in range(0, WAITTIME):
                         car.stop(True)
-                        car.x -= 4
+                        #print(small_timer)
+                        #small_timer+=1
+                    '''
+                    #car.x -= 4
 
-                
+                car.detect(1) 
                 if car.accel_bool:                    
                     car.accelerate(1)
 
-
-            elif car.x < screen_width*0.13 and car.y > screen_height*0.02:
-
+            elif car.x < 155 and car.y > 90:
                 #see if car is on left side, will go vertical down
-                if car.drop_val == 1:
-                    if(int(car.y)==car.drop_y) or (int(car.y-1)==car.drop_y):   
-                        car.stop(True) 
-                        #passenger.top = False
-                        car.y+=4
-                        #^working dropoff no passenger spawn
-
-                #passenger drop-off testing goes below
-
-                        #for y in range(0, wait_list[random.randint(0,2)]):
-                        #car.now = pygame.time.get_ticks()
-
-                    
-                        #for x in range(0, )
-
-                
-                '''
-                if car.pass_bool:
-                    passenger = make_passenger(car.x-15, car.drop_y)
-                    passenger.initial = car.x
-                    passenger_list.append(passenger)
-                    passenger.passenger_value = car.drop_val
-                    car.pass_bool = False
-
-
-                '''
-
                 car.detect(2)
                 if car.accel_bool:
                     car.accelerate(2)
+
                                         
             #get rid of car if it crosses bottom line (memory management)
-            if car.y >(screen_height*0.85) and car.x<(screen_width/2):
+            if car.y >(SCREEN_HEIGHT - 50) and car.x<(200):
                 del car_list[0]
+                car_count+=1
+                #print("cars crossed:", car_count)
+                print("a")
+                #cars_compare = 100
+
+                #LOOK HERE ANICHAU
+                #currently, on every car pass, write time needed
+                #if writing time only for all 100 to pass, change cars_compare to 100
+                cars_compare = 100
+                if car_count == cars_compare:
+                    hundred_cars_bool = True
+                    done = True
+
                 #fail safe to check if car going down is still red - not dropped off = bug and need to be fixed
+                
+                '''
+                #will go through this later
                 if car.color == RED:
                     print("car not dropped off, something wrong")
                     done = True
-
+                '''
 
         #pedestrian logic
         for person in person_list:
@@ -438,7 +371,7 @@ def main():
             #person coming from top leg
             elif person.personal_value == 1:
                 for y in range(10,25):
-                    for person_sweep in range(-10, 10):
+                    for person_sweep in range(-4, 4):
                         person_infront = screen.get_at(((int(person.x) + person_sweep), int(person.y) -y))
                         if person_infront==RED or person_infront==GREEN or person_infront==YELLOW:
                             person.stop()
@@ -450,7 +383,7 @@ def main():
             #person coming from left leg
             elif person.personal_value == 2:
                 for x in range(10,20):
-                    for person_sweep in range(-10, 10):
+                    for person_sweep in range(-4, 4):
                         person_infront = screen.get_at((int(person.x)-x, int(person.y)+person_sweep))
                         
                         if person_infront==RED or person_infront==GREEN or person_infront==YELLOW:
@@ -460,15 +393,16 @@ def main():
                 if person_bool:
                     person.x -= PERSON_SPEED
 
-            
             #always check if person has walked PERSON_WALK pixels, then erase from list (memory management)
 
             #delete from left leg
             if((person.personal_value == 0) and (abs(person.initial - person.x) > PERSON_WALK)):
+                
                 del(person_list[0])
 
             #delete from top leg
             elif((person.personal_value == 1) and (abs(person.initial - person.y) > PERSON_WALK)):
+
                 del(person_list[0])
 
             #delete from right leg
@@ -476,64 +410,64 @@ def main():
                 del(person_list[0])
 
 
-        #exiting passenger logic
-        for passenger in passenger_list:
-            #need distance check to erase
-            
-            #top leg passenger
-            if passenger.top:
-                passenger.y += -PERSON_SPEED/2
-            '''
-            if passenger.y<(screen_height*0.02):
-                passenger.y += -PERSON_SPEED/2
-            '''
-
-            '''
-            if passenger.passenger_value == 0:
-                passenger.y += -PERSON_SPEED/2
-                if (abs(passenger.initial - passenger.y) > 60):
-                    del(passenger_list[0])
-                
-            #left leg passenger
-            elif passenger.passenger_value == 1:
-                passenger.x += -PERSON_SPEED/2
-                if(abs(passenger.initial - passenger.x) > 60):
-                    del(passenger_list[0])
-            '''
-
         # --- Drawing
         # Set the screen background
         screen.fill(WHITE)
         
-        road_width = 100
-        road_vert = screen_height*0.70
         #draw the road
-        #template: (x, y, width, height)
-        #top side of road
-        pygame.draw.rect(screen, BLACK, (screen_width*0.1, 100, screen_width*0.75, road_width))
-        #right side of road
-        pygame.draw.rect(screen, BLACK, ((screen_width*0.9-road_width), 100, road_width, road_vert))
-        #left side of road
-        pygame.draw.rect(screen, BLACK, (screen_width*0.1, 100, road_width, road_vert))
+        pygame.draw.rect(screen, BLACK, (100, 50, 1200, 100))
+        pygame.draw.rect(screen, BLACK, (1200, 50, 100, 700))
+        pygame.draw.rect(screen, BLACK, (100, 50, 100, 700))
         
         # draw all moving objects
+        
         for car in car_list:
             pygame.draw.circle(screen, car.color, [int(car.x), int(car.y)], CAR_SIZE)
 
         for person in person_list:
             pygame.draw.circle(screen, GREEN, [int(person.x), int(person.y)], 7)
-
-        for passenger in passenger_list:
-            pygame.draw.circle(screen, passenger.color, [int(passenger.x), int(passenger.y)], 8)
         
+ 
         # --- Wrap-up, limit to 60 frames per second
         clock.tick(60)
- 
+        my_seconds+=1
+        mytime = my_seconds/60
+        #print(int(mytime))
+        
         # update screen with newly drawn
         pygame.display.flip()
  
-    # Close everything 
+    # Close everything
     pygame.quit()
+    if hundred_cars_bool:
+        body_print = "\n\nseconds needed for all " + str(car_count) + " cars to pass: %0.2f seconds" % mytime
+        print(body_print)
+
+        real_time = strftime("%Y-%m-%d %H-%M-%S", gmtime())
+
+        #LOOK HERE ANICHAU
+
+        save_path = "/Users/jeffrey/Documents/Github/carsim/files/testing"
+        filestring = "rawdata- " + str(real_time) + ".txt"
+        complete_name = os.path.join(save_path, filestring)
+        
+        f = open(complete_name, "a+")
+        #f = open(complete_name, "w")
+        #f = open("text.txt", "a+")
+
+        #various hundred car trials below
+        #f.write("\nthis is first go")
+        f.write("Time Elapsed for 100 Cars, No Pedestrians")
+        #write in the date or something else as well
+        #f.write("Time Elapsed for 100 Cars, Randomly Spawned Pedestrians")
+        #f.write("Time Elapsed for 100 Cars, Pedestrians in 10 Second Intervals")
+        #f.write("Time Elapsed for 100 Cars, Pedestrians in 20 Second Intervals")
+        #f.write("Time Elapsed for 100 Cars, Pedestrians in 30 Second Intervals")
+
+        f.write(body_print)
+        f.write("\n\n=================================================")
+        f.close()
+
  
 if __name__ == "__main__":
     main()
